requirements.txt
google-cloud-aiplatform
google-cloud-storage
google-cloud-firestore
python-docx
docx2pdf
pymupdf
langchain
tqdm
openpyxl
matplotlib
PyYAML
pandas
ftfy